+++ OpenKh.Engine/Parsers/MdlxParser.cs
@@ -1,6 +1,4 @@
-﻿using OpenKh.Common;
-using OpenKh.Engine.Maths;
-using OpenKh.Engine.Parsers.Kddf2;
+using OpenKh.Common;
 using OpenKh.Kh2;
 using OpenKh.Ps2;
 using System;
@@ -8,9 +6,31 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Numerics;
+using System.Runtime.InteropServices;
 
 namespace OpenKh.Engine.Parsers
 {
+    public class CustomVertex
+    {
+        [StructLayout(LayoutKind.Sequential)]
+        public struct PositionColoredTextured
+        {
+            public float X, Y, Z;
+            public int Color;
+            public float Tu, Tv;
+
+            public PositionColoredTextured(Vector3 v, int clr, float tu, float tv)
+            {
+                X = v.X;
+                Y = v.Y;
+                Z = v.Z;
+                Color = clr;
+                Tu = tu;
+                Tv = tv;
+            }
+        }
+    }
+
     public class MeshDescriptor
     {
         public CustomVertex.PositionColoredTextured[] Vertices;
@@ -27,7 +47,7 @@ namespace OpenKh.Engine.Parsers
             {
                 MeshDescriptors = new Kkdf2MdlxParser(mdlx.SubModels.First())
                     .ProcessVerticesAndBuildModel(
-                        MdlxMatrixUtil.BuildTPoseMatrices(mdlx.SubModels.First(), Matrix4x4.Identity)
+                        BuildTPoseMatrices(mdlx.SubModels.First(), Matrix4x4.Identity)
                     );
             }
             else if (IsMap(mdlx))
@@ -124,6 +144,57 @@ namespace OpenKh.Engine.Parsers
                 TextureIndex = vifPacketDescriptor.TextureId,
                 IsOpaque = vifPacketDescriptor.IsTransparentFlag == 0,
             };
+        }
+
+        /// <summary>
+        /// Build intial T-pose matrices.
+        /// </summary>
+        private static Matrix4x4[] BuildTPoseMatrices(Mdlx.SubModel model, Matrix4x4 initialMatrix)
+        {
+            var boneList = model.Bones.ToArray();
+            var matrices = new Matrix4x4[boneList.Length];
+            {
+                var absTranslationList = new Vector3[matrices.Length];
+                var absRotationList = new Quaternion[matrices.Length];
+                for (int x = 0; x < matrices.Length; x++)
+                {
+                    Quaternion absRotation;
+                    Vector3 absTranslation;
+                    var oneBone = boneList[x];
+                    var parent = oneBone.Parent;
+                    if (parent < 0)
+                    {
+                        absRotation = Quaternion.Identity;
+                        absTranslation = Vector3.Zero;
+                    }
+                    else
+                    {
+                        absRotation = absRotationList[parent];
+                        absTranslation = absTranslationList[parent];
+                    }
+
+                    var localTranslation = Vector3.Transform(new Vector3(oneBone.TranslationX, oneBone.TranslationY, oneBone.TranslationZ), Matrix4x4.CreateFromQuaternion(absRotation));
+                    absTranslationList[x] = absTranslation + localTranslation;
+
+                    var localRotation = Quaternion.Identity;
+                    if (oneBone.RotationZ != 0)
+                        localRotation *= (Quaternion.CreateFromAxisAngle(Vector3.UnitZ, oneBone.RotationZ));
+                    if (oneBone.RotationY != 0)
+                        localRotation *= (Quaternion.CreateFromAxisAngle(Vector3.UnitY, oneBone.RotationY));
+                    if (oneBone.RotationX != 0)
+                        localRotation *= (Quaternion.CreateFromAxisAngle(Vector3.UnitX, oneBone.RotationX));
+                    absRotationList[x] = absRotation * localRotation;
+                }
+                for (int x = 0; x < matrices.Length; x++)
+                {
+                    var absMatrix = initialMatrix;
+                    absMatrix *= Matrix4x4.CreateFromQuaternion(absRotationList[x]);
+                    absMatrix *= Matrix4x4.CreateTranslation(absTranslationList[x]);
+                    matrices[x] = absMatrix;
+                }
+            }
+
+            return matrices;
         }
     }
 }
