+++ OpenKh.Tests.Engine/MdlxLoaderTests.cs
@@ -1,4 +1,3 @@
-﻿using OpenKh.Engine.Parsers.Kddf2;
 using OpenKh.Ps2;
 using System.IO;
 using System.Linq;
